:sparkles: Séparation du fichier traverse
